+++ src/recon_lite/__init__.py
@@ -7,13 +7,17 @@ hierarchical planning and execution tasks.
 
 from .__version__ import __version__
 from .engine import ActivationMode, EngineConfig, ReConEngine
+from .formal_engine import EdgeMessage, FormalMessage, FormalReConEngine
 from .graph import Graph, LinkType, Node, NodeState, NodeType
 from .logger import RunLogger
 
 __all__ = [
     "__version__",
     "ActivationMode",
+    "EdgeMessage",
     "EngineConfig",
+    "FormalMessage",
+    "FormalReConEngine",
     "Graph",
     "LinkType",
     "Node",